Correction for short int conversion

+++ src/VME_BridgeVx718.cpp
@@ -121,7 +121,8 @@ namespace VME
   void
   BridgeVx718::InputRead(CVInputSelect input)
   {
-    short unsigned int data;
+    //short unsigned int data;
+    unsigned int data;
     if (CAENVME_ReadRegister(fHandle, cvInputReg, &data)!=cvSuccess) {
       std::ostringstream o;
       o << "Failed to read data input register #" << static_cast<int>(input);
